+++ credmark/cmf/credmark_dev.py
@@ -3,6 +3,7 @@ import sys
 import argparse
 import logging
 import json
+import unittest
 from typing import List, Union
 from importlib.metadata import version
 from dotenv import load_dotenv, find_dotenv
@@ -10,11 +11,12 @@ from dotenv import load_dotenv, find_dotenv
 sys.path.append('.')
 from .engine.context import EngineModelContext
 from .engine.model_loader import ModelLoader
-from .engine.mocks import ModelMockRunner
+from .engine.mocks import MockGenerator, ModelMockRunner
 from .engine.web3 import Web3Registry
 from .engine.model_api import ModelApi
-from credmark.dto import json_dump, json_dumps, print_example, print_tree, dto_schema_viz
-from credmark.dto.dto_error_schema import extract_error_codes_and_descriptions
+from credmark.dto import json_dump, json_dumps
+from credmark.cmf.engine.model_unittest import ModelTestContextFactory
+from credmark.cmf.model.print import print_manifest, print_manifest_description
 
 
 logger = logging.getLogger(__name__)
@@ -29,9 +31,7 @@ def add_api_url_arg(parser):
                         'You do not normally need to set this.')
 
 
-def main():
-    load_dotenv(find_dotenv(usecwd=True))
-
+def main():  # pylint: disable=too-many-statements
     parser = argparse.ArgumentParser(
         formatter_class=argparse.RawDescriptionHelpFormatter,
         description='Credmark developer tool')
@@ -102,20 +102,36 @@ def main():
         '-l', '--use_local_models', default=None,
         help='Comma-separated list of model slugs for models that should '
         'favor use of the local version. This is only required when a model is '
-        'calling another model.')
+        'calling another model. Use "*" to favor the use of local versions of all models.')
     parser_run.add_argument(
         '-m', '--model_mocks', default=None,
         help='Module path and symbol of model mocks config to use. '
         'For example, models.contrib.mymodels.mymocks.mock_config')
+    parser_run.add_argument(
+        '--generate_mocks', default=None,
+        help='Generate model mocks and write them to the specified file. '
+        'The generated python file can be used with --model_mocks on another run or in unit tests.')
     parser_run.add_argument('--provider_url_map', required=False, default=None,
                             help='JSON object of chain id to Web3 provider HTTP URL. '
-                            'Overrides settings in env vars.')
+                            'Overrides settings in env var or .env file.')
     add_api_url_arg(parser_run)
     parser_run.add_argument('--run_id', help=argparse.SUPPRESS, required=False, default=None)
     parser_run.add_argument('--depth', help=argparse.SUPPRESS, type=int, required=False, default=0)
-    parser_run.add_argument('model-slug', default='(missing model-slug arg)',
-                            help='Slug for the model to run.')
+    parser_run.add_argument(
+        'model-slug', default='(missing model-slug arg)',
+        help='Slug for the model to run or "console" for the interactive console.')
     parser_run.set_defaults(func=run_model, depth=0)
+
+    parser_test = subparsers.add_parser('test', help='Run model tests', aliases=['run-tests'])
+    parser_test.add_argument('-p', '--pattern', required=False, default='test*.py',
+                             help='Pattern to match test files (test*.py default).')
+    add_api_url_arg(parser_test)
+    parser_test.add_argument('--provider_url_map', required=False, default=None,
+                             help='JSON object of chain id to Web3 provider HTTP URL. '
+                             'Overrides settings in env var or .env.test file.')
+    parser_test.add_argument('tests_folder', nargs='?', default='models',
+                             help='Folder to start discovery for tests. Defaults to "models".')
+    parser_test.set_defaults(func=run_tests)
 
     parser_build = subparsers.add_parser(
         'build', help='Build model manifest [Not required during development]',
@@ -131,6 +147,12 @@ def main():
         sys.exit(1)
 
     args = parser.parse_args()
+
+    if args.func == run_tests:  # pylint: disable=comparison-with-callable
+        load_dotenv(find_dotenv('.env.test', usecwd=True))
+    else:
+        load_dotenv(find_dotenv(usecwd=True))
+
     args.func(vars(args))
 
 
@@ -150,6 +172,17 @@ def load_models(args, load_dev_models=False):
         model_paths = model_path.split(';')
     else:
         model_paths = None
+
+    # We add the tests folder (if set)
+    # to the models paths if not present
+    tests_folder = args.get('tests_folder')
+    if tests_folder:
+        if model_paths is None:
+            model_paths = [tests_folder]
+        else:
+            if tests_folder not in model_paths:
+                model_paths.append(tests_folder)
+
     model_loader = ModelLoader(model_paths, manifest_file, load_dev_models)
     model_loader.log_errors()
     return model_loader
@@ -301,63 +334,10 @@ def list_deployed_models(args):  # pylint: disable=too-many-branches
 
 def print_manifests(manifests: List[dict], describe_schemas=False):
     for m in manifests:  # pylint: disable=too-many-nested-blocks
-        for i, v in m.items():
-            if i == 'slug':
-                sys.stdout.write(f'{v}\n')
-                sys.stdout.write(f' - {i}: {v}\n')
-            else:
-                if not describe_schemas:
-                    sys.stdout.write(f' - {i}: {v}\n')
-                else:
-                    if i == 'input':
-                        input_tree = dto_schema_viz(
-                            v, v.get('title', 'Object'), v, 0, 'tree',
-                            only_required=False, tag='top', limit=10)
-                        input_examples = dto_schema_viz(
-                            v, v.get('title', 'Object'), v, 0, 'example',
-                            only_required=False, tag='top', limit=10)
-
-                        print(' - input schema (* for required field):')
-                        print_tree(input_tree, '   ', sys.stdout.write)
-
-                        print(' - input example:')
-                        print_example(input_examples, '   ', sys.stdout.write)
-
-                    elif i == 'output':
-                        output_tree = dto_schema_viz(
-                            v, v.get('title', 'Object'), v, 0, 'tree',
-                            only_required=False, tag='top', limit=1)
-                        output_examples = dto_schema_viz(
-                            v, v.get('title', 'Object'), v, 0, 'example',
-                            only_required=True, tag='top', limit=1)
-
-                        print(' - output schema (* for required field):')
-                        print_tree(output_tree, '   ', sys.stdout.write)
-
-                        print(' - output example:')
-                        print_example(output_examples, '   ', sys.stdout.write)
-
-                    elif i == 'error':
-                        codes = extract_error_codes_and_descriptions(v)
-                        print(' - errors:')
-                        if len(codes) > 0:
-                            for ct in codes:
-                                print(f'   {ct[0]}')
-                                print(f'     codes={ct[1]}')
-                                print(f'     {ct[2]}')
-                            title = v.get('title', 'Error')
-                            output_tree = dto_schema_viz(
-                                v, title, v, 0, 'tree', only_required=False, tag='top', limit=1)
-                            output_examples = dto_schema_viz(
-                                v, title, v, 0, 'example', only_required=False, tag='top', limit=1)
-                            print(' - error schema:')
-                            print_tree(output_tree, '   ', sys.stdout.write)
-                        else:
-                            print('   No defined errors')
-
-                    else:
-                        sys.stdout.write(f' - {i}: {v}\n')
-
+        if describe_schemas:
+            print_manifest_description(m, sys.stdout)
+        else:
+            print_manifest(m, sys.stdout)
         sys.stdout.write('\n')
 
 
@@ -397,6 +377,41 @@ def remove_manifest_file(args):
     sys.exit(0)
 
 
+def create_chain_to_provider_url(providers_json):
+    # if arg is provided, it overrides any providers env vars
+    if providers_json is not None:
+        try:
+            chain_to_provider_url = json.loads(providers_json)
+        except Exception as err:
+            logger.error(f'Error parsing JSON in arg provider_url_map: {err}')
+            raise
+    else:
+        try:
+            chain_to_provider_url = Web3Registry.load_providers_from_env()
+        except Exception as err:
+            logger.error(f'{err}')
+            raise
+    return chain_to_provider_url
+
+
+def run_tests(args):
+    config_logging(args, 'INFO')
+
+    api_url: Union[str, None] = args['api_url']
+    providers_json = args['provider_url_map']
+    chain_to_provider_url = create_chain_to_provider_url(providers_json)
+    model_loader = load_models(args, load_dev_models=True)
+
+    factory = ModelTestContextFactory(model_loader, chain_to_provider_url, api_url)
+    ModelTestContextFactory.use_factory(factory)
+
+    pattern = args['pattern']
+    tests_folder = args['tests_folder']
+    test_argv = [sys.argv[0], 'discover',
+                 '-s', tests_folder, '-p', pattern]
+    unittest.main(module=None, argv=test_argv, verbosity=2)
+
+
 def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,too-many-locals
     exit_code = 0
 
@@ -406,17 +421,7 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
         chain_to_provider_url = None
         providers_json = args['provider_url_map']
 
-        # if arg is provided, it overrides any providers env vars
-        if providers_json is not None:
-            try:
-                chain_to_provider_url = json.loads(providers_json)
-            except Exception as err:
-                logger.error(f'Error parsing JSON in arg provider_url_map: {err}')
-        else:
-            try:
-                chain_to_provider_url = Web3Registry.load_providers_from_env()
-            except Exception as err:
-                logger.error(f'{err}')
+        chain_to_provider_url = create_chain_to_provider_url(providers_json)
 
         model_loader = load_models(args, load_dev_models=not args.get(
             'run_id'))  # we assume developer will not pass a run_id
@@ -432,6 +437,7 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
         debug_log: bool = args['debug']
         use_local_models: Union[str, None] = args['use_local_models']
         model_mocks_config: Union[str, None] = args['model_mocks']
+        generate_mocks: Union[str, None] = args.get('generate_mocks')
 
         if use_local_models is not None and len(use_local_models):
             local_model_slugs = use_local_models.split(',')
@@ -441,6 +447,14 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
         if model_mocks_config:
             model_mock_runner = ModelMockRunner(model_mocks_config)
             EngineModelContext.use_model_mock_runner(model_mock_runner)
+
+        if generate_mocks is not None:
+            if not generate_mocks.endswith('.py'):
+                generate_mocks += '.py'
+            mock_gen = MockGenerator()
+            EngineModelContext.add_model_run_listener(mock_gen.model_run)
+        else:
+            mock_gen = None
 
         if debug_log:
             dbg_logger = logging.getLogger('credmark.cmf.engine.context.debug')
@@ -468,6 +482,10 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
             run_id=run_id,
             depth=depth)
 
+        if generate_mocks is not None and mock_gen is not None:
+            logger.info(f'Writing mocks to file "{generate_mocks}"')
+            mock_gen.write(generate_mocks, model_slug)
+
         if 'error' in result:
             etype = result.get('error', {}).get('type')
             if etype == 'ModelInputError':
@@ -479,10 +497,11 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
             else:
                 exit_code = 1
 
-        if format_json:
-            print(json_dumps(result, indent=4).replace('\\n', '\n').replace('\\"', '\''))
-        else:
-            json_dump(result, sys.stdout)
+        if model_slug != 'console':
+            if format_json:
+                print(json_dumps(result, indent=4).replace('\\n', '\n').replace('\\"', '\''))
+            else:
+                json_dump(result, sys.stdout)
 
     except Exception as e:
         # this exception would only happen have been raised
@@ -494,7 +513,8 @@ def run_model(args):  # pylint: disable=too-many-statements,too-many-branches,to
                 "message": f'Error in credmark-dev: {str(e)}'
             }
         }
-        json.dump(msg, sys.stdout)
+        if args.get('model-slug') != 'console':
+            json.dump(msg, sys.stdout)
         exit_code = 1
     finally:
         sys.stdout.write('\n')