+++ HISTORY.rst
@@ -3,5 +3,20 @@
 History
 -------
 
+0.8.17 [2022-05-12]
+* Model console
+
+0.8.16 [2022-05-10]
+* Model mock generation
+
+0.8.15 [2022-05-09]
+* chainId and blockNumber in model run results
+
+0.8.14 [2022-05-09]
+* Unit testing support
+
+* 0.8.13 [2022-05-03]
+Contract ledger queries
+
 0.0.1 [2022-02-25]
 * First public release